refactor env.go
